refactor(vrops-tests): Moved to correct place Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
