+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/operations/configuration/ConfigurationVropsTest.java
@@ -12,13 +12,13 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.configuration;
+package com.vmware.pscoe.iac.artifact.aria.operations.configuration;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import com.vmware.pscoe.iac.artifact.aria.operations.configuration.ConfigurationVrops;
+import com.vmware.pscoe.iac.artifact.configuration.ConfigurationException;
 
 import java.net.UnknownHostException;
 import java.util.Properties;
